Remove getTimeByteWidth, replace with const 8
I realized that the only places calling getTimeByteWidth were the
Timestamp and Duration blocks, which are the two blocks that always use
BigInt64Array. So rather than keep the function call for those two
scenarios, which should (if I understand correctly) always return 8
anyway, I just dropped in the constant.