Use const 8 byteWidth for Timestamp and Duration (#130)
* Remove getTimeByteWidth, replace with const 8

I realized that the only places calling getTimeByteWidth were the
Timestamp and Duration blocks, which are the two blocks that always use
BigInt64Array. So rather than keep the function call for those two
scenarios, which should (if I understand correctly) always return 8
anyway, I just dropped in the constant.

* yarn.lock and package.json changes

I wasn't sure if I should commit these or not, but this is what I had to
do in onder to get yarn test to run locally.

* Bump yarn version for Volta CI

Per advice on PR, bumping the CI yarn version to match my local version,
4.9.4

* Ran yarn fmt to fix CI

Found the prettier command in the package.json scripts so I ran it :)

* Replace ArrayBuffer with ArrayBufferLike

* Uncomment & fix timestamp test

Per PR request, uncommented the "timestamp" test in ffi.test.ts and then
recommented the console.log calls (as I assume they're unwanted with a
succeeding test). Also "fixed" the test by replacing the word "test"
with "it", which seemed more in-line with the rest of the file.